src/engine/pyproject.toml
[project]
name = "ovd-engine"
version = "0.1.0"
description = "OVD Platform — Motor LangGraph multi-agente"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.3.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "psycopg[binary,pool]>=3.2.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
    "nats-py>=2.6.0",
    "duckduckgo-search>=6.0.0",
    # Sprint 9 — Secrets Management (Infisical)
    "infisical-python>=1.5.0",
    # Sprint 10 — Auth: JWT access tokens + refresh tokens
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2,bcrypt]>=1.7.4",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
markers = [
    "unit: test unitario sin infraestructura (default)",
    "integration: requiere PostgreSQL y NATS reales",
    "e2e: smoke test contra engine completo",
    "regression: verifica contratos de sprints anteriores",
]

[tool.uv]
# uv genera uv.lock automáticamente al hacer `uv sync`
# Commitear uv.lock para reproducibilidad en CI/CD y Docker
